Move verbose_git() and quiet_git() to mesonlib
